test(mobile): speed up NotificationsPromptStakeFlow suite by trimming render tree
Register only the stake flow under test per parametrized case instead of
all family flow screens. Suite time drops ~32s to ~9s locally with full
coverage retained (79/79 passing).

LIVE-31711
