+++ .changeset/swift-tests-trim.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": patch
+---
+
+Speed up NotificationsPromptStakeFlow integration tests by registering only the stake flow under test in each parametrized case instead of all ~37 family flow screens.
