+++ apps/ledger-live-mobile/src/mvvm/features/NotificationsPrompt/__integrations__/NotificationsPromptStakeFlow.part-1.test.tsx
@@ -1,0 +1,5 @@
+import { stakePromptCaseChunks } from "./stakePromptFixtures";
+import { describeStakePromptCaseChunk } from "./stakePromptCaseTests";
+
+// Each part runs a case chunk so the test runner can parallelize the stake prompt flow.
+describeStakePromptCaseChunk(stakePromptCaseChunks[0], 1);
